pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/graphql_to_csv"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[project]
name = "graphql_to_csv"
version = "0.0.1"
description = "Open Source GraphQL to CSV"
authors = [
  { name="Eric Wilson", email="eric.wilson@geekcafe.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "graphql-core"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
